requirements.txt
Django==5.2.1
django-crispy-forms==2.4
django-filter==25.1
django-simple-history==3.8.0
django-soft-delete==1.0.18
django-tables2==2.7.5
django-widget-tweaks==1.5.0
crispy-bootstrap5==2025.4
libsass==0.23.0
altcha==0.1.9
whitenoise==6.9.0
Brotli==1.1.0
JSON-log-formatter==1.1.1
celery==5.5.2
redis==5.3.0
django_celery_results==2.6.0
mysqlclient==2.2.7
python-decouple==3.8
colorama==0.4.6
